refactor&docs: extract module combing logic & wrtie annotation
