+++ bot/discord.js
@@ -2,7 +2,10 @@ const { BOT_TOKEN, CLIENT_ID, GUILD_ID, PROXY_URL } = require('./config.json')
 const { ignoreMessage } = require(
   './message/ignoreMessage/ignoreMessage.module')
 const { useProxy } = require('./useProxy/useProxy')
-const { useCommandsCreate } = require('./interaction/command/commandCreate')
+const { useCommandsCreate } = require('./interaction/command/commandCreate/commandCreate')
+const { commandReaction } = require('./interaction/command/commandReaction/commandReaction')
+const { autoCreateChannel } = require('./guild/channel/channelCreate/autoCreateChannel.module')
+const { messageReaction } = require('./message/messageReaction/messageReaction.module')
 
 //start using proxy
 useProxy(PROXY_URL)
@@ -19,39 +22,27 @@ client.on('ready', () => {
 })
 
 client.on('interactionCreate', async interaction => {
+  //it is should a command
   if (!interaction.isCommand()) return
-
-  if (interaction.commandName === 'ping') {
-    await interaction.reply('Pong!')
-  }
-
-  if (interaction.commandName === 'hello') {
-    await interaction.reply('Nice!')
-  }
+  //What should be done after the command is executed
+  commandReaction(interaction)
 })
 
 client.on('messageCreate', async (message) => {
+  //message that should be ignore
   ignoreMessage(message)
-  if (message.mentions.has(client.user.id)) {
-    await message.reply('Hello')
-  }
+  //The reaction to  particular messages
+  messageReaction(message)
 })
 
-client.on('guildCreate', async (server) =>{
+client.on('guildCreate', async (server) => {
   try {
-    await server.channels.create('may-be-scammer',{
-        type: 'GUILD_TEXT',
-        permissionOverwrites: [
-          {
-            id: server.id,
-            allow: ['VIEW_CHANNEL']
-          },
-        ],
-      }
-      )
+    //after the bot is deploy,automatic create a new channel
+    await autoCreateChannel(server)
   }
   catch (err) {
     console.log(err)
   }
 })
+
 client.login(BOT_TOKEN)
