+++ bot/message/messageReaction/messageReaction.module.js
@@ -1,0 +1,7 @@
+const messageReactionModule = (message) => {
+  if (message.mentions.has(message.client.user.id)) {
+    message.reply('Hello')
+  }
+}
+
+module.exports = { messageReaction: messageReactionModule }
